table_sum_float returns float now. (typeing mistake)
table_set_mixed calls dll correctly.(spelling mistake)
TableAggregate test method enhanced , looking for a float bug
TableDynamic example finished and passes tests
moved a bit further towards accepting IEnnumerable instead of object[], allowing for users to throw almost any kind of enmerable structure of enumerable collection objects into a subtable.
Added a few more methods that take a string instead of a column index.

+++ native/tightdb_c_cs/tightdb_c_cs.cpp
@@ -800,10 +800,11 @@ TIGHTDB_C_CS_API int64_t table_sum_int(Table * table_ptr , size_t column_ndx)
 {   
     return table_ptr->sum(column_ndx);
 }
-TIGHTDB_C_CS_API double table_sum_float(Table * table_ptr , size_t column_ndx)
+TIGHTDB_C_CS_API float table_sum_float(Table * table_ptr , size_t column_ndx)
 {   
     return table_ptr->sum_float(column_ndx);
 }
+
 TIGHTDB_C_CS_API double table_sum_double(Table * table_ptr , size_t column_ndx)
 {   
     return table_ptr->sum_double(column_ndx);
@@ -883,7 +884,7 @@ TIGHTDB_C_CS_API int64_t tableview_sum_int(TableView * tableview_ptr , size_t co
 {   
     return tableview_ptr->sum(column_ndx);
 }
-TIGHTDB_C_CS_API double tableview_sum_float(TableView * tableview_ptr , size_t column_ndx)
+TIGHTDB_C_CS_API float tableview_sum_float(TableView * tableview_ptr , size_t column_ndx)
 {   
     return tableview_ptr->sum_float(column_ndx);
 }
